# docs/swagger.yaml
definitions:
  modelsViewApi.BadRequestError:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  modelsViewApi.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  modelsViewApi.FinishInput:
    properties:
      falseStartList:
        description: Массив номеров команд в порядке финиша
        items:
          type: integer
        type: array
    type: object
...
  modelsViewApi.RaceFormData:
    properties:
      class:
        type: string
      date:
        type: string
      id:
        type: string
      number:
        type: integer
      rating-id:
        type: string
    type: object
...
  modelsViewApi.RaceInput:
    properties:
      class:
        type: integer
      date:
        type: string
      number:
        type: integer
    required:
    - date
    - number
    type: object
...
  modelsViewApi.StartInput:
    properties:
      falseStartList:
        description: Массив номеров фальш-стартовавших команд
        items:
          type: integer
        type: array
      specCircumstance:
        description: Номер специального обстоятельства в случае фальш-старта
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/ratings/{ratingID}/races:
    get:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      responses:
        "200":
          description: Список гонок успешно получен
          schema:
            items:
              $ref: '#/definitions/modelsViewApi.RaceFormData'
            type: array
        "404":
          description: Рейтинг или гонки не найдены
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить гонки по ID рейтинга
      tags:
      - Race
    post:
      consumes:
      - application/json
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Входные данные для создания гонки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.RaceInput'
      produces:
      - application/json
      responses:
        "201":
          description: Гонка успешно создана
          schema:
            $ref: '#/definitions/modelsViewApi.RaceInput'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Рейтинг не найден
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Создать новую гонку для рейтинга
      tags:
      - Race
  /api/ratings/{ratingID}/races/{raceID}:
    delete:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки, которую нужно удалить
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Гонка успешно удалена
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Гонка не найдена
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Удалить гонку
      tags:
      - Race
    get:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о гонке успешно получена
          schema:
            $ref: '#/definitions/modelsViewApi.RaceFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Гонка не найдена
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Получить информацию о гонке
      tags:
      - Race
    put:
      consumes:
      - application/json
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Входные данные для обновления гонки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.RaceInput'
      produces:
      - application/json
      responses:
        "200":
          description: Гонка успешно обновлена
          schema:
            $ref: '#/definitions/modelsViewApi.RaceFormData'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Гонка не найдена
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Обновить информацию о гонке
      tags:
      - Race
  /api/ratings/{ratingID}/races/{raceID}/finish:
    post:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Входные данные для завершения финишной процедуры
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.FinishInput'
      produces:
      - application/json
      responses:
        "200":
          description: Процедура финиша успешно выполнена
          schema:
            type: string
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Гонка не найдена
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Завершить финишную процедуру гонки
      tags:
      - Race
...
  /api/ratings/{ratingID}/races/{raceID}/start:
    post:
      parameters:
      - description: Уникальный идентификатор рейтинга
        format: uuid
        in: path
        name: ratingID
        required: true
        type: string
      - description: Уникальный идентификатор гонки
        format: uuid
        in: path
        name: raceID
        required: true
        type: string
      - description: Входные данные для завершения стартовой процедуры
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/modelsViewApi.StartInput'
      produces:
      - application/json
      responses:
        "200":
          description: Процедура старта успешно выполнена
          schema:
            type: string
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/modelsViewApi.BadRequestError'
        "404":
          description: Гонка не найдена
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/modelsViewApi.ErrorResponse'
      summary: Завершить стартовую процедуру гонки
      tags:
      - Race
...
swagger: "2.0"